Created BoxBall Class
Created a new BoxBall class. I copy and pasted the BouncingBall class as a starting off point.
Added 3 local variables to serve as walls for the box.
Added in xSpeed to serve as lateral acceleration.
Modified move() to test if the ball has hit any of the sides of the box.