more on boxBounce()
I've been fiddling quite a bit, so I'm not sure what the last official change I made was.
I'm having a bit of trouble figuring out how to get the bounce loop to end.
I added a condition test for the boxBounce parameter.
I also moved the HashSet initialization to inside the boxBounce method so that each time it's called, the old set goes away.
I modified the move method in the BoxBounce class a bit and I added more accessor methods to use in the Ball Demo class. I might not need those, though, so they might get axed.