Specifically ending the bouncing.
The balls move okay, but I'm having trouble ending the method.
I was thinking that I'd have a counter that increased each time a ball was done moving and once that number was equal to the size of the set the bounce loop would end, but since the condition to test if a ball is finished is inside a for each loop, the same ball that's been done for multiple iterations artificially increases the finishedBalls counter. so the method ends before all the balls are done.
I thought about removing the boxBall object from the set once it's done, but then the next iteration of the for loop throws an error.