Defeat. Sad, miserable defeat
I can't figure out how to end the boxBounce method elegantly, so a regular for loop it is. The balls will each get 300 iterations of their move method and then the loop terminates.
This makes me sad :( I tried so hard to find another way to do it so that the loop terminated as soon as all the balls were done moving.