Slight modifications
Adjusted some values in boxBall class to make the balls a bit more animated and more likely to bounce off multiple walls.
Adjusted canvas wait time so animation looks okay regardless of the amount of balls in the box.
The color of the box changes every time you call the boxBounce method, and I've no idea how to fix that.