Extra Credit and no Gravity
So, as it turns out... no gravity needed.
I removed the gravity and ballDegradation fields from BoxBall class and their subsequent uses.
Also fixed documentation.
Added canvasBounce() to BallDemo which is the extra credit. The box is the dimensions of the canvas.
I added a constructor to BallDemo that takes a width and height to set up a new canvas to better utilize the canvasBounce() method.
With that, I think, Week 7 Lab is complete and final.